requirements.txt
# 🎯 AI-Powered Career Guidance Platform - Updated Requirements
# Resolved dependency conflicts (October 2025)

# Core Framework
flask>=3.0.0
flask-cors>=4.0.0
flasgger>=0.9.7


# LangChain Components (Compatible versions)
langchain>=0.2.0,<0.3.0
langchain-google-genai>=1.0.0,<2.0.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0

# Google AI Libraries (Compatible versions)
google-generativeai>=0.7.0,<0.8.0

# API and Environment Management
python-dotenv>=1.0.0
google-search-results>=2.4.2

# File Processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Additional Utilities
requests>=2.31.0
urllib3>=2.0.0

